+++ feature_engine/discretisation/gemoetric_width.py
